Add Collection class and use extend for Popover
